Added First Version of Projectile and Crossbow Scripts
Crossbow script follows mouse cursor and rotates the crossbow in the player's hand to follow the cursor. It also allows shooting projectiles with a certain cooldown between shots (chosen in inspector). Projectile script actually makes the arrows move and destroys them after a certain amount of rime (chosen in inspector). Made room for a destroy effect for when the arrows disappear, and made an arrow prefab.