Updated Projectile Script for Destruction Effects
I updated the projectile script to now include 2 particle effects as game objects; one for an arrow hitting the wall and the other for an arrow hitting an enemy. I updated the DestroyProjectile function to take in booleans and run the current destruciton effect accordingly (or none at all if necessary). I made a seperate function for DestroyProjectile to work with invoke