Finalized Crossbow Charging Logic and Sprite Changing
Crossbow charging is now polished and working. Added sprite changing to the charging logic. If the mouse if held for half of the full charge time, it switches to a half charged sprite. If it is released before fully charging, it returns to the no charge sprite. If the mouse is held for the duration of the full charge time, crossbow enters charge state and stays in that state until the user clicks again to shoot an arrow.