Made Rock Throwing Script
I made the rock both a sprite and a projectile prefab. I then made a rock script allowing the rock to act like a weapon, rotates with mouse, and can be fired consecutively as long as the time between shots is sufficient. Uses the projectile script to handle the actual movement of the rock